Cleaned up log scan implementation
