+++ cs/src/core/Allocator/BlittableAllocator.cs
@@ -15,55 +15,6 @@ using System.Diagnostics;
 
 namespace FASTER.core
 {
-    internal class Frame : IDisposable
-    {
-        public readonly int frameSize, pageSize, sectorSize;
-        public readonly byte[][] frame;
-        public GCHandle[] handles;
-        public long[] pointers;
-
-        public Frame(int frameSize, int pageSize, int sectorSize)
-        {
-            this.frameSize = frameSize;
-            this.pageSize = pageSize;
-            this.sectorSize = sectorSize;
-
-            frame = new byte[frameSize][];
-            handles = new GCHandle[frameSize];
-            pointers = new long[frameSize];
-        }
-        public void Allocate(int index)
-        {
-            var adjustedSize = pageSize + 2 * sectorSize;
-            byte[] tmp = new byte[adjustedSize];
-            Array.Clear(tmp, 0, adjustedSize);
-
-            handles[index] = GCHandle.Alloc(tmp, GCHandleType.Pinned);
-            long p = (long)handles[index].AddrOfPinnedObject();
-            pointers[index] = (p + (sectorSize - 1)) & ~(sectorSize - 1);
-            frame[index] = tmp;
-        }
-
-        public void Clear(int pageIndex)
-        {
-            Array.Clear(frame[pageIndex], 0, frame[pageIndex].Length);
-        }
-
-        public long GetPhysicalAddress(long frameNumber, long offset)
-        {
-            return pointers[frameNumber % frameSize] + offset;
-        }
-
-        public void Dispose()
-        {
-            for (int i = 0; i < frameSize; i++)
-            {
-                handles[i].Free();
-                frame[i] = null;
-                pointers[i] = 0;
-            }
-        }
-    }
     public unsafe sealed class BlittableAllocator<Key, Value> : AllocatorBase<Key, Value>
         where Key : new()
         where Value : new()
@@ -342,7 +293,7 @@ namespace FASTER.core
         /// <returns></returns>
         public override IFasterScanIterator<Key, Value> Scan(long beginAddress, long endAddress, ScanBufferingMode scanBufferingMode)
         {
-            return new BlittableScanIterator(this, beginAddress, endAddress, scanBufferingMode);
+            return new BlittableScanIterator<Key, Value>(this, beginAddress, endAddress, scanBufferingMode);
         }
 
 
@@ -359,12 +310,12 @@ namespace FASTER.core
         /// <param name="devicePageOffset"></param>
         /// <param name="device"></param>
         /// <param name="objectLogDevice"></param>
-        private void AsyncReadPagesFromDeviceToFrame<TContext>(
+        internal void AsyncReadPagesFromDeviceToFrame<TContext>(
                                         long readPageStart,
                                         int numPages,
                                         IOCompletionCallback callback,
                                         TContext context,
-                                        Frame frame,
+                                        BlittableFrame frame,
                                         out CountdownEvent completed,
                                         long devicePageOffset = 0,
                                         IDevice device = null, IDevice objectLogDevice = null)
@@ -406,160 +357,7 @@ namespace FASTER.core
                 usedDevice.ReadAsync(offsetInFile, (IntPtr)frame.pointers[pageIndex], (uint)AlignedPageSizeBytes, callback, asyncResult);
             }
         }
-
-        /// <summary>
-        /// Scan iterator for hybrid log
-        /// </summary>
-        public class BlittableScanIterator : IFasterScanIterator<Key, Value>
-        {
-            private readonly int frameSize;
-
-            private readonly BlittableAllocator<Key, Value> hlog;
-            private readonly long beginAddress, endAddress;
-
-            private readonly Frame frame;
-            private readonly CountdownEvent[] loaded;
-
-            private bool first = true;
-            private long currentAddress;
-
-            /// <summary>
-            /// Constructor
-            /// </summary>
-            /// <param name="hlog"></param>
-            /// <param name="beginAddress"></param>
-            /// <param name="endAddress"></param>
-            /// <param name="scanBufferingMode"></param>
-            public BlittableScanIterator(BlittableAllocator<Key, Value> hlog, long beginAddress, long endAddress, ScanBufferingMode scanBufferingMode)
-            {
-                this.hlog = hlog;
-                this.beginAddress = beginAddress;
-                this.endAddress = endAddress;
-                currentAddress = beginAddress;
-
-                if (scanBufferingMode == ScanBufferingMode.SinglePageBuffering)
-                    frameSize = 1;
-                else
-                    frameSize = 2;
-
-                frame = new Frame(frameSize, hlog.PageSize, hlog.sectorSize);
-                loaded = new CountdownEvent[frameSize];
-
-                var frameNumber = (currentAddress >> hlog.LogPageSizeBits) % frameSize;
-                hlog.AsyncReadPagesFromDeviceToFrame
-                    (currentAddress >> hlog.LogPageSizeBits,
-                    1, AsyncReadPagesCallback, Empty.Default,
-                    frame, out loaded[frameNumber]);
-            }
-
-
-            public bool GetNext(out Key key, out Value value)
-            {
-                key = default(Key);
-                value = default(Value);
-
-                while (true)
-                {
-                    // Check for boundary conditions
-                    if (currentAddress >= endAddress)
-                    {
-                        return false;
-                    }
-
-                    if (currentAddress < hlog.BeginAddress)
-                    {
-                        throw new Exception("Iterator address is less than log BeginAddress " + hlog.BeginAddress);
-                    }
-
-                    var currentPage = currentAddress >> hlog.LogPageSizeBits;
-                    var currentFrame = currentPage % frameSize;
-                    var offset = currentAddress & hlog.PageSizeMask;
-
-                    if (currentAddress < hlog.HeadAddress)
-                        BufferAndLoad(currentAddress, currentPage, currentFrame);
-
-                    // Check if record fits on page, if not skip to next page
-                    if ((currentAddress & hlog.PageSizeMask) + recordSize > hlog.PageSize)
-                    {
-                        currentAddress = (1 + (currentAddress >> hlog.LogPageSizeBits)) << hlog.LogPageSizeBits;
-                        continue;
-                    }
-
-
-                    if (currentAddress >= hlog.HeadAddress)
-                    {
-                        // Read record from cached page memory
-                        var _physicalAddress = hlog.GetPhysicalAddress(currentAddress);
-
-                        key = hlog.GetKey(_physicalAddress);
-                        value = hlog.GetValue(_physicalAddress);
-                        currentAddress += hlog.GetRecordSize(_physicalAddress);
-                        return true;
-                    }
-
-                    var physicalAddress = frame.GetPhysicalAddress(currentFrame, offset);
-                    key = hlog.GetKey(physicalAddress);
-                    value = hlog.GetValue(physicalAddress);
-                    currentAddress += hlog.GetRecordSize(physicalAddress);
-                    return true;
-                }
-            }
-
-            
-            private void BufferAndLoad(long currentAddress, long currentPage, long currentFrame)
-            {
-                if (first || (currentAddress & hlog.PageSizeMask) == 0)
-                {
-                    // Prefetch pages based on buffering mode
-                    if (frameSize == 1)
-                    {
-                        if (!first)
-                        {
-                            hlog.AsyncReadPagesFromDeviceToFrame(currentAddress >> hlog.LogPageSizeBits, 1, AsyncReadPagesCallback, Empty.Default, frame, out loaded[currentFrame]);
-                        }
-                    }
-                    else
-                    {
-                        var endPage = endAddress >> hlog.LogPageSizeBits;
-                        if ((endPage > currentPage) &&
-                            ((endPage > currentPage + 1) || ((endAddress & hlog.PageSizeMask) != 0)))
-                        {
-                            hlog.AsyncReadPagesFromDeviceToFrame(1 + (currentAddress >> hlog.LogPageSizeBits), 1, AsyncReadPagesCallback, Empty.Default, frame, out loaded[(currentPage + 1) % frameSize]);
-                        }
-                    }
-                    first = false;
-                }
-                loaded[currentFrame].Wait();
-            }
-
-            public void Dispose()
-            {
-                throw new NotImplementedException();
-            }
-
-            private void AsyncReadPagesCallback(uint errorCode, uint numBytes, NativeOverlapped* overlap)
-            {
-                if (errorCode != 0)
-                {
-                    Trace.TraceError("OverlappedStream GetQueuedCompletionStatus error: {0}", errorCode);
-                }
-
-                var result = (PageAsyncReadResult<Empty>)Overlapped.Unpack(overlap).AsyncResult;
-
-                if (result.freeBuffer1.buffer != null)
-                {
-                    hlog.PopulatePage(result.freeBuffer1.GetValidPointer(), result.freeBuffer1.required_bytes, result.page);
-                    result.freeBuffer1.Return();
-                }
-
-                if (result.handle != null)
-                {
-                    result.handle.Signal();
-                }
-
-                Interlocked.MemoryBarrier();
-                Overlapped.Free(overlap);
-            }
-        }
     }
 }
+
+
